logmsg-serialize-fixup: try to reuse the same NVHandle when fixing up handles
In the deserialization path, we always looked up the new handle for the
name being deserialized. This is not always needed, especially when the
same syslog-ng instance is reading back the message (which often happens
in the case of a disk queue).

This patch whether the old handle still references the same name, which
we can do without going through the NVRegistry hash table. (e.g.
O(1) instead of O(logN)). Seems to increase deserialization speed from
150k to 180k/sec, at least with -O2. Interestingly it gets slower
without O2.

But now as we do have information about wheter any of the handles changed
at all, we can also remove a lot of processing, namely:
  - updating of the SDATA array
  - sorting the index

This boosts performance from 180k to 270k/sec.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>