nvtable: rename NVDynValue to NVIndexEntry to describe its purpose better
An NVDynValue is confusing in a number of ways:
  * NVTable is storing a list of name-value pairs, so NVDynValue is what,
    especially as NVEntry is the struct that describes a specific
    name-value pair
  * the concept of dynamic vs. static entries is not always clear as we
    are talking about an unsorted list of name-value pairs
  * we sometimes use dyn_value in other cases dyn_entries when referring
    to the same thing.

So this patch attempts to clarify naming as a first step by renaming
NVDynValue to NVIndexEntry, as it is the struct that is one element in
the lookup table that is used to look up name-value pairs by handle.

So in essence, the dynamic-entries array is an index that facilitates
handle based lookup.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>